+++ D.hpp
@@ -1,0 +1,6 @@
+struct Particulas{
+  double x,y;// posición del la particula
+  double Vx,Vy;//velocidad
+  double F;//fuerza sobre ella
+  double M=1;//masa  1(provicional)
+};
